+++ js/object-creation.js
@@ -1,0 +1,6 @@
+import {createPhotoDescription} from './photo-description';
+
+// Создание полного объекта
+const similarPhotoDescriptions = Array.from({length: 25}, createPhotoDescription);
+
+export {similarPhotoDescriptions};
